move logic to useTrade hook
